整合gzk
